refactor: Move parser and lexer to core
