+++ src/core/parser/lexer.rs
@@ -1,9 +1,6 @@
-pub mod tests;
-pub mod token;
-
 use std::{iter::Peekable, str::CharIndices};
 
-use crate::lexer::token::{KEYWORD_REGISTRY, Token, TokenKind};
+use crate::core::parser::token::{KEYWORD_REGISTRY, Token, TokenKind};
 
 #[derive(Clone, Debug)]
 pub struct Lexer<'src> {
@@ -28,7 +25,7 @@ impl<'src> Lexer<'src> {
     }
 
     fn next_token(&mut self) -> Token {
-        use super::lexer::token::TokenKind::*;
+        use crate::core::parser::token::TokenKind::*;
         let registry = &KEYWORD_REGISTRY;
         self.skip_whitespaces();
