+++ src/evaluator/tests.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 mod evaluator_tests {
     use crate::{
+        core::parser::Parser,
         evaluator::{self, context::Context, object::Object},
-        parser::Parser,
     };
 
     #[test]
