+++ src/tests/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(test)]
+pub mod lexer_tests;
+
+#[cfg(test)]
+pub mod parser_tests;
